+++ addon/components/flexberry-groupedit.js
@@ -188,7 +188,7 @@ export default FlexberryBaseComponent.extend({
   mainModelProjection: undefined,
 
   /**
-    Flag: indicates whether ordering by clicking on column headers is allowed.
+    Flag indicates whether ordering by clicking on column headers is allowed.
 
     @example
     If sorting is used then there has to be declaration:
@@ -217,7 +217,7 @@ export default FlexberryBaseComponent.extend({
   placeholder: t('components.flexberry-groupedit.placeholder'),
 
   /**
-    Flag: indicates whether table rows are clickable (action will be fired after row click).
+    Flag indicates whether table rows are clickable (action will be fired after row click).
 
     @property rowClickable
     @type Boolean
@@ -226,7 +226,7 @@ export default FlexberryBaseComponent.extend({
   rowClickable: false,
 
   /**
-    Flag: indicates whether to save current model before going to the detail's route.
+    Flag indicates whether to save current model before going to the detail's route.
 
     @example
     This flag is used when flag {{#crossLink "FlexberryGroupeditComponent/editOnSeparateRoute:property"}}{{/crossLink}} is enabled.