+++ addon/mixins/flexberry-objectlistview-route.js
@@ -23,7 +23,8 @@ export default Ember.Mixin.create({
       @param {Ember.Object} record Record related to clicked table row
     */
     objectListViewRowClick(record, editFormRoute) {
-      this.transitionTo(editFormRoute, record.get('id'));
+      let recordId = record.get('id') || record.get('data.id');
+      this.transitionTo(editFormRoute, recordId);
     },
 
     /**